CRC: move traits to lib.rs
